Move synthesis rules into monsters package
