+++ monsters/synthesis_rules.py
